[fix] 챌린지 삭제 챌린지 오너만 삭제 할 수 있게 변경 (어드민 허용 )

+++ src/main/kotlin/com/meokq/api/challenge/service/ChallengeService.kt
@@ -137,8 +137,9 @@ class ChallengeService(
     fun delete(challengeId: String, authReq: AuthReq) {
         val challenge = findModelById(challengeId)
         checkNotNull(challenge.status)
-        challenge.status.deleteAction()
-
+        if(challenge.customerId != authReq.userId && authReq.userType == UserType.CUSTOMER){
+            challenge.status.deleteAction()
+        }
         questRepository.findById(challenge.questId!!)
             .ifPresent {
                 it.rewards?.let {
